backend/requirements.txt
# core framework
fastapi==0.110.0
uvicorn[standard]==0.29.0

# database / models
sqlmodel==0.0.16
sqlalchemy==2.0.30   # explicit for stability

# templates
jinja2==3.1.4

# multipart/form-data (file uploads, form posts)
python-multipart==0.0.9

# auth + security
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
email-validator==2.1.1  # needed for pydantic[email]

# testing/dev
pytest==8.2.0
httpx==0.27.0
